Move spring_mass_system from drake/examples to drake/systems/framework/examples.
