+++ drake/systems/framework/examples/CMakeLists.txt
@@ -1,0 +1,7 @@
+# This directory has no library targets. All the source code here is part
+# drakeSystemFramework, which is declared in systems/framework.
+
+drake_install_headers(
+  spring_mass_system.h)
+
+add_subdirectory(test)
