+++ drake/systems/framework/examples/test/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_executable(spring_mass_system_test spring_mass_system_test.cc)
-target_link_libraries(spring_mass_system_test drakeSpringMassSystem
-        drakeSystemFramework ${GTEST_BOTH_LIBRARIES})
+target_link_libraries(spring_mass_system_test drakeSystemFramework
+    ${GTEST_BOTH_LIBRARIES})
 add_test(NAME spring_mass_system_test COMMAND spring_mass_system_test)
